server/requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
asyncio==3.4.3
attrs==21.2.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
frozenlist==1.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==5.2.0
pycares==4.1.2
pycparser==2.21
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
